requirements.txt
# POLYHEAL AI – Backend Requirements
# Install with: pip install -r requirements.txt

# API Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# AI / LLM
openai>=1.20.0
google-generativeai>=0.4.0

# Vector Store (optional — used by advanced memory features)
faiss-cpu>=1.7.4
chromadb>=0.4.0

# HTTP & Web
httpx>=0.27.0
requests>=2.31.0

# Env management
python-dotenv>=1.0.0

# Data
pydantic>=2.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
